requirements.txt
python-dotenv==1.0.1
google-cloud-storage==3.1.0
google-cloud-aiplatform==1.83.0
